Enable MobileAP only when backhaul connection is required
Currently MobileAP is enabled in constructor of LocNetIface
and never disabled. Since location_hal_daemon now launches
the location framework including LocNetIface, so MobileAP is
always on. So enable MobileAP only when backhaul connection
is required and disable MobileAP after we disconnect the
backhaul connection.

For station mode status indications, now we make use of the
new function RegisterForIndications in QCMAP. There is no
need now to enable MobileAP just for Backhaul or network
availability notification. We need to use this API for
targets that support MobileAP RegisterForIndications.

Connect Request count need to incremented only after state
of backhaul moves to CONNECT.

Change-Id: I068769bdd60d749247892ad9e848ca66a03b7268
CRs-Fixed: 2559583