fixing measurement registration
measurements can only be registered at the first time LocApiV02::open
is called. The following opens may mistakenly deregister it if gnss
poly is part of the registration mask. This is never an issue for
android but a problem for automotive.

Fixed it by moving the possible config call after each time reg
mask is updated.

Change-Id: I65c5452c929d4a02337ff283fc371c9b779a4914
CRs-Fixed: 2725132