added QRTR status watcher to LocIpcQrtrRecver
LocIpcQrtrRecver is able to monitor up / down of specified
services, as well as client going away. The reason it has
to be a recver for such events is that sender doesn't have
a listening socket.

In addition, LocationApiService no longer keep track of
QRTR clients with files in the filesystem. LCA / LIA are
now using their own watchers to keep an eye on the UP event
of LocationApiService.

Change-Id: I176873962011cdb62c8ccf00844c9038609df50d
CRs-Fixed: 2616587